Updated folder names & added generic error message
