+++ src/components/editMeasure/measureDetails/MeasureDetailsSidebar.tsx
