+++ src/components/editMeasure/measureDetails/__mocks__/MeasureDetails.tsx
